requirements.txt
constructs>=10.0.0,<11.0.0
python-dotenv==1.0.0
google-genai==1.2.0
tenacity==9.0.0
requests==2.31.0
beautifulsoup4==4.12.3
httpx==0.27.2
tqdm
arxiv
praw==7.7.1
feedparser==6.0.10
regex==2023.12.25